Set feglm Control Parameters

Set and change parameters used for fitting feglm. Termination conditions are similar to glm.

Usage

feglm_control(
  dev_tol = 1e-06,
  center_tol = 1e-06,
  iter_max = 25L,
  iter_center_max = 10000L,
  iter_inner_max = 50L,
  iter_interrupt = 1000L,
  iter_ssr = 10L,
  limit = 10L,
  trace = FALSE,
  drop_pc = TRUE,
  keep_mx = FALSE
)

Arguments

dev_tol

tolerance level for the first stopping condition of the maximization routine. The stopping condition is based on the relative change of the deviance in iteration \(r\) and can be expressed as follows: \(|dev_{r} - dev_{r - 1}| / (0.1 + |dev_{r}|) < tol\). The default is 1.0e-08.

center_tol

tolerance level for the stopping condition of the centering algorithm. The stopping condition is based on the relative change of the centered variable similar to the 'lfe' package. The default is 1.0e-08.

iter_max

unsigned integer indicating the maximum number of iterations in the maximization routine. The default is 25L.

iter_center_max

unsigned integer indicating the maximum number of iterations in the centering algorithm. The default is 10000L.

iter_inner_max

unsigned integer indicating the maximum number of iterations in the inner loop of the centering algorithm. The default is 50L.

iter_interrupt

unsigned integer indicating the maximum number of iterations before the algorithm is interrupted. The default is 1000L.

iter_ssr

unsigned integer indicating the number of iterations to skip before checking if the sum of squared residuals improves. The default is 10L.

limit

unsigned integer indicating the maximum number of iterations of theta.ml. The default is 10L.

trace

logical indicating if output should be produced in each iteration. Default is FALSE.

drop_pc

logical indicating to drop observations that are perfectly classified/separated and hence do not contribute to the log-likelihood. This option is useful to reduce the computational costs of the maximization problem and improves the numerical stability of the algorithm. Note that dropping perfectly separated observations does not affect the estimates. The default is TRUE.

keep_mx

logical indicating if the centered regressor matrix should be stored. The centered regressor matrix is required for some covariance estimators, bias corrections, and average partial effects. This option saves some computation time at the cost of memory. The default is TRUE.

Value

A named list of control parameters.

See also

Examples

feglm_control(0.05, 0.05, 10L, 10L, TRUE, TRUE, TRUE)
#> $dev_tol
#> [1] 0.05
#> 
#> $center_tol
#> [1] 0.05
#> 
#> $iter_max
#> [1] 10
#> 
#> $iter_center_max
#> [1] 10
#> 
#> $iter_inner_max
#> [1] 1
#> 
#> $iter_interrupt
#> [1] 1
#> 
#> $iter_ssr
#> [1] 1
#> 
#> $limit
#> [1] 10
#> 
#> $trace
#> [1] FALSE
#> 
#> $drop_pc
#> [1] TRUE
#> 
#> $keep_mx
#> [1] FALSE
#>