Home 🏠

Functions apes bias_corr feglm felm fenegbin fepoisson fit_control summary_table

Vignettes Poisson Pseudo-Maximum Likelihood (PPML) Model with Cluster-Robust Standard Errors separation

News Changelog

felm {capybara}	R Documentation

LM fitting with high-dimensional k-way fixed effects

Description

feglm can be used to fit linear models with many high-dimensional fixed effects. The estimation procedure is based on unconditional maximum likelihood and can be interpreted as a “weighted demeaning” approach.

Usage

felm(formula = NULL, data = NULL, weights = NULL, control = NULL)

Arguments

formula 	 an object of class "formula": a symbolic description of the model to be fitted. formula must be of type response ~ slopes | fixed_effects | cluster.

data 	 an object of class "data.frame" containing the variables in the model. The expected input is a dataset with the variables specified in formula and a number of rows at least equal to the number of variables in the model.

weights 	 an optional string with the name of the prior weights variable in data.

control 	 a named list of parameters for controlling the fitting process. See fit_control for details.

Value

A named list of class "felm". The list contains the following eleven elements:

coefficients 	 a named vector of the estimated coefficients

fitted_values 	 a vector of the estimated dependent variable

weights 	 a vector of the weights used in the estimation

hessian 	 a matrix with the numerical second derivatives

null_deviance 	 the null deviance of the model

nobs 	 a named vector with the number of observations used in the estimation indicating the dropped and perfectly predicted observations

fe_levels 	 a named vector with the number of levels in each fixed effect

nms_fe 	 a list with the names of the fixed effects variables

formula 	 the formula used in the model

data 	 the data used in the model after dropping non-contributing observations

control 	 the control list used in the model

References

Gaure, S. (2013). "OLS with Multiple High Dimensional Category Variables". Computational Statistics and Data Analysis, 66.

Marschner, I. (2011). "glm2: Fitting generalized linear models with convergence problems". The R Journal, 3(2).

Stammann, A., F. Heiss, and D. McFadden (2016). "Estimating Fixed Effects Logit Models with Large Panel Data". Working paper.

Stammann, A. (2018). "Fast and Feasible Estimation of Generalized Linear Models with High-Dimensional k-Way Fixed Effects". ArXiv e-prints.

Examples

# Model with fixed effects
mod <- felm(log(mpg) ~ log(wt) | cyl, mtcars)
summary(mod)

# Model without fixed effects but with clustered standard errors
# Note: Use 0 to indicate no fixed effects when specifying clusters
mod <- felm(log(mpg) ~ log(wt) | 0 | cyl, mtcars)
summary(mod)