Home 🏠

Functions apes bias_corr feglm felm fenegbin fepoisson fit_control summary_table

Vignettes Poisson Pseudo-Maximum Likelihood (PPML) Model with Cluster-Robust Standard Errors Nonexistence of estimates of Poisson models

News Changelog

summary_table {capybara}	R Documentation

Generate formatted regression tables

Description

Generate formatted regression tables

Usage

summary_table(
  ...,
  coef_digits = 3,
  se_digits = 3,
  stars = TRUE,
  latex = FALSE,
  model_names = NULL,
  caption = NULL,
  label = NULL
)

Arguments

... 	 One or more model objects of felm or feglm class.

coef_digits 	 Number of digits for coefficients. The default is 3.

se_digits 	 Number of digits for standard errors. The default is 3.

stars 	 Whether to include significance stars. The default is TRUE.

latex 	 Whether to output as LaTeX code. The default is FALSE.

model_names 	 Optional vector of custom model names

caption 	 Optional caption for the table (LaTeX only)

label 	 Optional label for cross-referencing (LaTeX only)

Value

A formatted table

Examples

m1 <- felm(mpg ~ wt | cyl, mtcars)
m2 <- fepoisson(mpg ~ wt | cyl, mtcars)
summary_table(m1, m2, model_names = c("Linear", "Poisson"))