capybara Documentation

Home 🏠

Functions fe_table feglm felm fenegbin fepoisson_asymmetric fepoisson fit_control sandwich_vcov summary_table

Vignettes Asymmetric Poisson Pseudo-Maximum Likelihood (APPML) Poisson Pseudo-Maximum Likelihood (PPML) Model with Cluster-Robust Standard Errors Nonexistence of estimates of Poisson models Different Variance-Covariance Estimators

News Changelog

vcov.feglm {capybara}	R Documentation

Covariance matrix for GLMs

Description

Covariance matrix for the estimator of the structural parameters from objects returned by feglm. The covariance is computed during model fitting - either the inverse Hessian (default) or the sandwich estimator if cluster variables are specified in the formula.

Usage

## S3 method for class 'feglm'
vcov(object, ...)

Arguments

object 	 an object of class "feglm".

... 	 additional arguments (currently ignored).

Value

A named matrix of covariance estimates.

See Also

feglm

Examples

# Model with clustering - returns sandwich covariance
ross2004_s1 <- ross2004[ross2004$year == 1994, ]
ross2004_s1 <- ross2004_s1[ross2004_s1$ltrade >
  quantile(ross2004_s1$ltrade, 0.75), ]

ross2004_s2 <- ross2004[ross2004$year == 1999, ]
ross2004_s2 <- ross2004_s2[ross2004_s2$ltrade >
  quantile(ross2004_s2$ltrade, 0.75), ]

ross2004_subset <- rbind(ross2004_s1, ross2004_s2)

fit <- fepoisson(ltrade ~ ldist + border | ctry1 | year, ross2004_subset)

vcov(fit)