Broom Integration

The provided broom methods do the following:

augment: Takes the input data and adds additional columns with the fitted values and residuals.

glance: Extracts the deviance, null deviance, and the number of observations.`

tidy: Extracts the estimated coefficients and their standard errors.

Usage

# S3 method for class 'feglm'
augment(x, newdata = NULL, ...)

# S3 method for class 'felm'
augment(x, newdata = NULL, ...)

# S3 method for class 'feglm'
glance(x, ...)

# S3 method for class 'felm'
glance(x, ...)

# S3 method for class 'feglm'
tidy(x, conf_int = FALSE, conf_level = 0.95, ...)

# S3 method for class 'felm'
tidy(x, conf_int = FALSE, conf_level = 0.95, ...)

Arguments

x

A fitted model object.

newdata

Optional argument to use data different from the data used to fit the model.

...

Additional arguments passed to the method.

conf_int

Logical indicating whether to include the confidence interval.

conf_level

The confidence level for the confidence interval.

Value

A tibble with the respective information for the augment, glance, and tidy methods.

Examples

mod <- fepoisson(mpg ~ wt | cyl, mtcars)
broom::augment(mod)
#> Registered S3 methods overwritten by 'broom':
#>   method       from    
#>   augment.felm capybara
#>   glance.felm  capybara
#>   tidy.felm    capybara
#> # A tibble: 32 × 5
#>      mpg    wt cyl   .fitted .residuals
#>    <dbl> <dbl> <fct>   <dbl>      <dbl>
#>  1  21    2.62 6        21.6    -0.552 
#>  2  21    2.88 6        20.6     0.415 
#>  3  22.8  2.32 4        26.4    -3.58  
#>  4  21.4  3.22 6        19.4     2.04  
#>  5  18.7  3.44 8        16.6     2.14  
#>  6  18.1  3.46 6        18.5    -0.429 
#>  7  14.3  3.57 8        16.2    -1.88  
#>  8  24.4  3.19 4        22.6     1.85  
#>  9  22.8  3.15 4        22.7     0.0828
#> 10  19.2  3.44 6        18.6     0.604 
#> # ℹ 22 more rows
broom::glance(mod)
#> # A tibble: 1 × 6
#>   deviance null_deviance nobs_full nobs_na nobs_pc  nobs
#> *    <dbl>         <dbl>     <int>   <int>   <int> <int>
#> 1     7.52          54.5        32       0       0    32
broom::tidy(mod)
#> # A tibble: 1 × 4
#>   estimate std.error statistic p.value
#> *    <dbl>     <dbl>     <dbl>   <dbl>
#> 1   -0.180    0.0716     -2.51  0.0120