LM fitting with high-dimensional k-way fixed effects

A wrapper for feglm with family = gaussian().

felm(formula = NULL, data = NULL, weights = NULL)

Arguments

formula

an object of class "formula": a symbolic description of the model to be fitted. formula must be of type y ~ x | k, where the second part of the formula refers to factors to be concentrated out. It is also possible to pass clustering variables to feglm as y ~ x | k | c.

data

an object of class "data.frame" containing the variables in the model.

weights

an optional string with the name of the 'prior weights' variable in data.

Value

The function felm returns a named list of class

"felm".

References

Gaure, S. (2013). "OLS with Multiple High Dimensional Category Variables". Computational Statistics and Data Analysis, 66.

Marschner, I. (2011). "glm2: Fitting generalized linear models with convergence problems". The R Journal, 3(2).

Stammann, A., F. Heiss, and D. McFadden (2016). "Estimating Fixed Effects Logit Models with Large Panel Data". Working paper.

Stammann, A. (2018). "Fast and Feasible Estimation of Generalized Linear Models with High-Dimensional k-Way Fixed Effects". ArXiv e-prints.

Examples

mod <- felm(
  log(trade) ~ log_dist + lang + cntg + clny | exp_year + imp_year,
  trade_panel
)

summary(mod)
#> Formula: log(trade) ~ log_dist + lang + cntg + clny | exp_year + imp_year
#> <environment: 0x5e9310805ba0>
#> 
#> Estimates:
#> 
#> |          | Estimate  | Std. Error | z value     | Pr(>|z|)   |
#> |----------|-----------|------------|-------------|------------|
#> | log_dist | -723.5816 |     0.0108 | -67305.7879 | 0.0000 *** |
#> | lang     |  -57.4220 |     0.0223 |  -2579.8225 | 0.0000 *** |
#> | cntg     | 6949.3930 |     0.0445 | 156240.6350 | 0.0000 *** |
#> | clny     | -744.1953 |     0.0447 | -16649.5540 | 0.0000 *** |
#> 
#> Significance codes: *** 99.9%; ** 99%; * 95%; . 90%
#> 
#> R-squared     : 0.2761 
#> Adj. R-squared: 0.2541 
#> 
#> Number of observations: Full 28152; Missing 0; Perfect classification 0