Negative Binomial model fitting with high-dimensional k-way fixed effects

A routine that uses the same internals as feglm.

fenegbin(
  formula = NULL,
  data = NULL,
  weights = NULL,
  beta.start = NULL,
  eta.start = NULL,
  init.theta = NULL,
  link = c("log", "identity", "sqrt"),
  control = NULL
)

Arguments

formula

an object of class "formula": a symbolic description of the model to be fitted. formula must be of type y ~ x | k, where the second part of the formula refers to factors to be concentrated out. It is also possible to pass clustering variables to feglm as y ~ x | k | c.

data

an object of class "data.frame" containing the variables in the model.

weights

an optional string with the name of the 'prior weights' variable in data.

beta.start

an optional vector of starting values for the structural parameters in the linear predictor. Default is \(\boldsymbol{\beta} = \mathbf{0}\).

eta.start

an optional vector of starting values for the linear predictor.

init.theta

an optional initial value for the theta parameter (see glm.nb).

link

the link function. Must be one of "log", "sqrt", or "identity".

control

a named list of parameters for controlling the fitting process. See feglm_control for details.

Examples

# same as the example in fepoisson but with overdispersion/underdispersion
mod <- fenegbin(
  trade ~ dist + lang + cntg + clny | exp_year + imp_year,
  trade_panel
)

summary(mod)
#> Warning: subscript out of bounds, not all 'rhs' available
#> Formula: trade ~ dist + lang + cntg + clny | exp_year + imp_year
#> <environment: 0x64516a1942c0>
#> 
#> Family: Negative Binomial(0.3514)
#> 
#> Estimates:
#> 
#> |      | Estimate | Std. error | z value   | Pr(> |z|)  |
#> |------|----------|------------|-----------|------------|
#> | dist |  -0.0005 |     0.0000 | -163.2177 | 0.0000 *** |
#> | lang |  -0.7636 |     0.0382 |  -19.9643 | 0.0000 *** |
#> | cntg |  -0.5745 |     0.0704 |   -8.1648 | 0.0000 *** |
#> | clny |   1.1913 |     0.0755 |   15.7843 | 0.0000 *** |
#> 
#> Significance codes: *** 99.9%; ** 99%; * 95%; . 90%
#> 
#> Number of observations: Full 28566; Missing 0; Perfect classification 0 
#> 
#> Number of Fisher Scoring iterations: 1
#> Number of outer iterations: 5
#> theta= 0.3514, std. error= 0.002743