Home 🏠

Functions apes bias_corr feglm felm fenegbin fepoisson fit_control summary_table

Vignettes Poisson Pseudo-Maximum Likelihood (PPML) Model with Cluster-Robust Standard Errors separation

News Changelog

fepoisson {capybara}	R Documentation

Poisson model fitting high-dimensional with k-way fixed effects

Description

A wrapper for feglm with family = poisson().

Usage

fepoisson(
  formula = NULL,
  data = NULL,
  weights = NULL,
  beta_start = NULL,
  eta_start = NULL,
  offset = NULL,
  control = NULL
)

Arguments

formula 	 an object of class "formula": a symbolic description of the model to be fitted. formula must be of type response ~ slopes | fixed_effects | cluster.

data 	 an object of class "data.frame" containing the variables in the model. The expected input is a dataset with the variables specified in formula and a number of rows at least equal to the number of variables in the model.

weights 	 an optional string with the name of the prior weights variable in data.

beta_start 	 an optional vector of starting values for the structural parameters in the linear predictor. Default is \boldsymbol{\beta} = \mathbf{0}.

eta_start 	 an optional vector of starting values for the linear predictor.

offset 	 an optional formula or numeric vector specifying an a priori known component to be included in the linear predictor. If a formula, it should be of the form ~ log(variable).

control 	 a named list of parameters for controlling the fitting process. See fit_control for details.

Value

A named list of class "feglm".

Examples

# check the feglm examples for the details about clustered standard errors
mod <- fepoisson(mpg ~ wt | cyl, mtcars)
summary(mod)