Set feglm Control Parameters

Set and change parameters used for fitting feglm. Termination conditions are similar to glm.

Usage

fit_control(
  dev_tol = 1e-08,
  center_tol = 1e-08,
  collin_tol = 1e-10,
  step_halving_factor = 0.5,
  alpha_tol = 1e-08,
  iter_max = 25L,
  iter_center_max = 10000L,
  iter_inner_max = 50L,
  iter_alpha_max = 10000L,
  iter_interrupt = 1000L,
  iter_ssr = 10L,
  step_halving_memory = 0.9,
  max_step_halving = 2L,
  start_inner_tol = 1e-06,
  accel_start = 6L,
  project_tol_factor = 0.001,
  grand_accel_tol = 1e-10,
  project_group_tol = 1e-12,
  irons_tuck_tol = 1e-10,
  grand_accel_interval = 5L,
  irons_tuck_interval = 3L,
  ssr_check_interval = 40L,
  convergence_factor = 1.1,
  tol_multiplier = 20,
  return_fe = TRUE,
  keep_tx = FALSE,
  init_theta = 0
)

Arguments

dev_tol

tolerance level for the first stopping condition of the maximization routine. The stopping condition is based on the relative change of the deviance in iteration \(r\) and can be expressed as follows: \(|dev_{r} - dev_{r - 1}| / (0.1 + |dev_{r}|) < tol\). The default is 1.0e-08.

center_tol

tolerance level for the stopping condition of the centering algorithm. The stopping condition is based on the relative change of the centered variable similar to the 'lfe' package. The default is 1.0e-08.

collin_tol

tolerance level for detecting collinearity. The default is 1.0e-07.

step_halving_factor

numeric indicating the factor by which the step size is halved to iterate towards convergence. This is used to control the step size during optimization. The default is 0.5.

alpha_tol

tolerance for fixed effects (alpha) convergence. The default is 1.0e-06.

iter_max

unsigned integer indicating the maximum number of iterations in the maximization routine. The default is 25L.

iter_center_max

unsigned integer indicating the maximum number of iterations in the centering algorithm. The default is 10000L.

iter_inner_max

unsigned integer indicating the maximum number of iterations in the inner loop of the centering algorithm. The default is 50L.

iter_alpha_max

maximum iterations for fixed effects computation. The default is 10000L.

iter_interrupt

unsigned integer indicating the maximum number of iterations before the algorithm is interrupted. The default is 1000L.

iter_ssr

unsigned integer indicating the number of iterations to skip before checking if the sum of squared residuals improves. The default is 10L.

step_halving_memory

numeric memory factor for step-halving algorithm. Controls how much of the previous iteration is retained. The default is 0.9.

max_step_halving

maximum number of post-convergence step-halving attempts. The default is 2.

start_inner_tol

starting tolerance for inner solver iterations. The default is 1.0e-04.

accel_start

Integer. Iteration to start conjugate gradient acceleration in centering. The default is 6L.

project_tol_factor

Factor to multiply center_tol for projection tolerance. The default is 1e-3.

grand_accel_tol

Tolerance for grand acceleration convergence. The default is 1e-10.

project_group_tol

Tolerance for individual group projections. The default is 1e-12.

irons_tuck_tol

Tolerance for Irons-Tuck acceleration. The default is 1e-10.

grand_accel_interval

Interval for applying grand acceleration. The default is 5L.

irons_tuck_interval

Interval for applying Irons-Tuck acceleration. The default is 3L.

ssr_check_interval

Interval for adaptive SSR convergence checks. The default is 40L.

convergence_factor

Factor for detecting slow convergence. The default is 1.1.

tol_multiplier

Multiplier for early termination tolerance check. The default is 20.0.

return_fe

logical indicating if the fixed effects should be returned. This can be useful when fitting general equilibrium models where skipping the fixed effects for intermediate steps speeds up computation. The default is TRUE and only applies to the feglm class.

keep_tx

logical indicating if the centered regressor matrix should be stored. The centered regressor matrix is required for some covariance estimators, bias corrections, and average partial effects. This option saves some computation time at the cost of memory. The default is TRUE.

init_theta

Initial value for the negative binomial dispersion parameter (theta). The default is 0.0.

Value

A named list of control parameters.

See also

Examples

fit_control(0.05, 0.05, 10L, 10L, TRUE, TRUE, TRUE)
#> $dev_tol
#> [1] 0.05
#> 
#> $center_tol
#> [1] 0.05
#> 
#> $collin_tol
#> [1] 10
#> 
#> $step_halving_factor
#> [1] 10
#> 
#> $alpha_tol
#> [1] TRUE
#> 
#> $iter_max
#> [1] 1
#> 
#> $iter_center_max
#> [1] 1
#> 
#> $iter_inner_max
#> [1] 50
#> 
#> $iter_alpha_max
#> [1] 10000
#> 
#> $iter_interrupt
#> [1] 1000
#> 
#> $iter_ssr
#> [1] 10
#> 
#> $step_halving_memory
#> [1] 0.9
#> 
#> $max_step_halving
#> [1] 2
#> 
#> $start_inner_tol
#> [1] 1e-06
#> 
#> $accel_start
#> [1] 6
#> 
#> $project_tol_factor
#> [1] 0.001
#> 
#> $grand_accel_tol
#> [1] 1e-10
#> 
#> $project_group_tol
#> [1] 1e-12
#> 
#> $irons_tuck_tol
#> [1] 1e-10
#> 
#> $grand_accel_interval
#> [1] 5
#> 
#> $irons_tuck_interval
#> [1] 3
#> 
#> $ssr_check_interval
#> [1] 40
#> 
#> $convergence_factor
#> [1] 1.1
#> 
#> $tol_multiplier
#> [1] 20
#> 
#> $return_fe
#> [1] TRUE
#> 
#> $keep_tx
#> [1] FALSE
#> 
#> $init_theta
#> [1] 0
#>