Home 🏠

Functions apes bias_corr feglm felm fenegbin fepoisson fit_control summary_table

Vignettes Poisson Pseudo-Maximum Likelihood (PPML) Model with Cluster-Robust Standard Errors Nonexistence of estimates of Poisson models

News Changelog

fit_control {capybara}	R Documentation

Set feglm Control Parameters

Description

Set and change parameters used for fitting feglm, felm, and fenegbin. Termination conditions are similar to glm.

Usage

fit_control(
  dev_tol = 1e-08,
  center_tol = 1e-08,
  collin_tol = 1e-10,
  step_halving_factor = 0.5,
  alpha_tol = 1e-08,
  iter_max = 25L,
  iter_center_max = 10000L,
  iter_inner_max = 50L,
  iter_alpha_max = 10000L,
  iter_interrupt = 1000L,
  step_halving_memory = 0.9,
  max_step_halving = 2L,
  start_inner_tol = 1e-06,
  return_fe = TRUE,
  keep_tx = FALSE,
  check_separation = TRUE,
  init_theta = 0
)

Arguments

dev_tol 	 tolerance level for the first stopping condition of the maximization routine. The stopping condition is based on the relative change of the deviance in iteration r and can be expressed as follows: |dev_{r} - dev_{r - 1}| / (0.1 + |dev_{r}|) < tol. The default is 1.0e-08.

center_tol 	 tolerance level for the stopping condition of the centering algorithm. The stopping condition is based on the relative change of the centered variable similar to the 'lfe' package. The default is 1.0e-08.

collin_tol 	 tolerance level for detecting collinearity. The default is 1.0e-07.

step_halving_factor 	 numeric indicating the factor by which the step size is halved to iterate towards convergence. This is used to control the step size during optimization. The default is 0.5.

alpha_tol 	 tolerance for fixed effects (alpha) convergence. The default is 1.0e-06.

iter_max 	 integer indicating the maximum number of iterations in the maximization routine. The default is 25L.

iter_center_max 	 integer indicating the maximum number of iterations in the centering algorithm. The default is 10000L.

iter_inner_max 	 integer indicating the maximum number of iterations in the inner loop of the centering algorithm. The default is 50L.

iter_alpha_max 	 maximum iterations for fixed effects computation. The default is 10000L.

iter_interrupt 	 integer indicating the maximum number of iterations before the algorithm is interrupted. The default is 1000L.

step_halving_memory 	 numeric memory factor for step-halving algorithm. Controls how much of the previous iteration is retained. The default is 0.9.

max_step_halving 	 maximum number of post-convergence step-halving attempts. The default is 2.

start_inner_tol 	 starting tolerance for inner solver iterations. The default is 1.0e-04.

return_fe 	 logical indicating if the fixed effects should be returned. This can be useful when fitting general equilibrium models where skipping the fixed effects for intermediate steps speeds up computation. The default is TRUE and only applies to the feglm class.

keep_tx 	 logical indicating if the centered regressor matrix should be stored. The centered regressor matrix is required for some covariance estimators, bias corrections, and average partial effects. This option saves some computation time at the cost of memory. The default is TRUE.

check_separation 	 logical indicating whether to perform separation detection for Poisson models. When TRUE (default), observations with perfect prediction are automatically detected and excluded from estimation. Set to FALSE to skip this check and speed up computation when separation is known not to be an issue. The default is TRUE.

init_theta 	 Initial value for the negative binomial dispersion parameter (theta). The default is 0.0.

Value

A named list of control parameters.

See Also

feglm, felm, and fenegbin

Examples

fit_control(0.05, 0.05, 10L, 10L, TRUE, TRUE, TRUE)