Recover the estimates of the fixed effects after fitting (G)LMs

The system might not have a unique solution since we do not take collinearity into account. If the solution is not unique, an estimable function has to be applied to our solution to get meaningful estimates of the fixed effects.

fixed_effects(object = NULL, alpha_tol = 1e-08)

Arguments

object

an object of class "feglm".

alpha_tol

tolerance level for the stopping condition. The algorithm is stopped at iteration \(i\) if \(||\boldsymbol{\alpha}_{i} - \boldsymbol{\alpha}_{i - 1}||_{2} < tol ||\boldsymbol{\alpha}_{i - 1}|| {2}\). Default is 1.0e-08.

Value

A named list containing named vectors of estimated fixed effects.

References

Stammann, A. (2018). "Fast and Feasible Estimation of Generalized Linear Models with High-Dimensional k-way Fixed Effects". ArXiv e-prints.

Gaure, S. (n. d.). "Multicollinearity, identification, and estimable functions". Unpublished.

See also

Examples

# check the feglm examples for the details about clustered standard errors

# subset trade flows to avoid fitting time warnings during check
set.seed(123)
trade_2006 <- trade_panel[trade_panel$year == 2006, ]
trade_2006 <- trade_2006[sample(nrow(trade_2006), 1000), ]

mod <- fepoisson(
  trade ~ log_dist + lang + cntg + clny | exp_year + imp_year,
  trade_2006
)

fixed_effects(mod)
#> $exp_year
#>          exp_year
#> ARG2006 12.120993
#> AUS2006 12.763306
#> AUT2006 12.256833
#> BEL2006 13.319330
#> BGR2006 10.655277
#> BOL2006  8.966858
#> BRA2006 13.906807
#> CAN2006 13.047891
#> CHE2006 12.304737
#> CHL2006 12.316111
#> CHN2006 15.237263
#> CMR2006  8.093573
#> COL2006 11.345813
#> CRI2006 10.661125
#> CYP2006  7.930249
#> DEU2006 14.587122
#> DNK2006 12.038110
#> ECU2006 10.527013
#> EGY2006 10.460228
#> ESP2006 13.507246
#> FIN2006 12.183546
#> FRA2006 14.035988
#> GBR2006 13.438160
#> GRC2006 11.106394
#> HKG2006 12.148183
#> HUN2006 11.815420
#> IDN2006 12.970274
#> IND2006 13.046334
#> IRL2006 12.284992
#> IRN2006  9.918407
#> ISL2006  7.085654
#> ISR2006 11.805084
#> ITA2006 13.582801
#> JOR2006  7.898916
#> JPN2006 14.876461
#> KEN2006  7.896458
#> KOR2006 14.287133
#> KWT2006  9.477292
#> LKA2006  8.658249
#> MAC2006  7.931897
#> MAR2006 10.563161
#> MEX2006 12.986382
#> MLT2006  9.385121
#> MMR2006  8.800165
#> MUS2006 10.370378
#> MWI2006  7.795832
#> MYS2006 13.096302
#> NER2006  6.246289
#> NGA2006  8.654755
#> NLD2006 13.321690
#> NOR2006 11.527787
#> NPL2006  6.474274
#> PAN2006  9.712261
#> PHL2006 12.816535
#> POL2006 11.725220
#> PRT2006 11.230132
#> QAT2006 10.472461
#> ROM2006 11.049452
#> SEN2006  6.141103
#> SGP2006 13.731638
#> SWE2006 12.895090
#> THA2006 12.958489
#> TTO2006  9.752789
#> TUN2006  9.629356
#> TUR2006 12.441507
#> TZA2006 10.050030
#> URY2006  9.879469
#> USA2006 15.199036
#> ZAF2006 12.619892
#> 
#> $imp_year
#>            imp_year
#> ARG2006  0.80127320
#> AUS2006  2.03027207
#> AUT2006  0.25549161
#> BEL2006  1.13931174
#> BGR2006 -0.70890847
#> BOL2006 -1.59167576
#> BRA2006  1.34953467
#> CAN2006  1.51728812
#> CHE2006  0.67649074
#> CHL2006  0.68731416
#> CHN2006  2.85520227
#> CMR2006 -2.44971287
#> COL2006  0.43305362
#> CRI2006 -0.54519090
#> CYP2006 -1.08723631
#> DEU2006  2.84789740
#> DNK2006 -0.17317876
#> ECU2006 -0.41066346
#> EGY2006 -0.59770798
#> ESP2006  1.69025535
#> FIN2006  0.42778773
#> FRA2006  2.13444597
#> GBR2006  2.09369162
#> GRC2006  0.10461380
#> HKG2006  1.55137303
#> HUN2006  0.13405577
#> IDN2006  0.66887254
#> IND2006  1.55468543
#> IRL2006 -0.10214578
#> IRN2006  0.51116994
#> ISL2006 -1.99169050
#> ISR2006  0.03528291
#> ITA2006  1.97913852
#> JOR2006 -1.54804821
#> JPN2006  2.29031815
#> KEN2006 -0.93915230
#> KOR2006  1.29959272
#> KWT2006 -0.73930835
#> LKA2006 -1.27231713
#> MAC2006 -3.06042374
#> MAR2006 -0.67783754
#> MEX2006  2.13285599
#> MLT2006 -1.90364504
#> MMR2006 -2.20524358
#> MUS2006 -1.99752277
#> MWI2006 -4.26262340
#> MYS2006  1.57633296
#> NER2006 -3.21682692
#> NGA2006 -0.02387191
#> NLD2006  1.83182019
#> NOR2006  0.32253008
#> NPL2006 -4.75805605
#> PAN2006 -1.06673928
#> PHL2006  0.65970630
#> POL2006  0.36144783
#> PRT2006  0.65448403
#> QAT2006 -0.34989473
#> ROM2006 -0.15378128
#> SEN2006 -2.70395326
#> SGP2006  1.78268119
#> SWE2006  0.89947619
#> THA2006  0.72531687
#> TTO2006 -2.03452116
#> TUN2006 -1.54799179
#> TUR2006  1.08998528
#> TZA2006 -1.92571719
#> URY2006 -1.97136591
#> USA2006  3.84025590
#> ZAF2006  1.11127995
#>