Recover the estimates of the fixed effects after fitting (G)LMs

The system might not have a unique solution since we do not take collinearity into account. If the solution is not unique, an estimable function has to be applied to our solution to get meaningful estimates of the fixed effects.

Usage

fixed_effects(object = NULL, control = NULL)

Arguments

object

an object of class "feglm".

control

a list of control parameters. If NULL, the default control parameters are used.

Value

A named list containing named vectors of estimated fixed effects.

References

Stammann, A. (2018). "Fast and Feasible Estimation of Generalized Linear Models with High-Dimensional k-way Fixed Effects". ArXiv e-prints.

Gaure, S. (n. d.). "Multicollinearity, identification, and estimable functions". Unpublished.

See also

Examples

# check the feglm examples for the details about clustered standard errors

# subset trade flows to avoid fitting time warnings during check
set.seed(123)
trade_2006 <- trade_panel[trade_panel$year == 2006, ]
trade_2006 <- trade_2006[sample(nrow(trade_2006), 500), ]

mod <- fepoisson(
  trade ~ log_dist + lang + cntg + clny | exp_year + imp_year,
  trade_2006
)

fixed_effects(mod)
#> $exp_year
#>          exp_year
#> ARG2006 12.337500
#> AUS2006 12.279329
#> AUT2006 11.443794
#> BEL2006 12.585595
#> BGR2006  9.938005
#> BOL2006  7.978908
#> BRA2006 13.178732
#> CAN2006 13.202071
#> CHE2006 11.488738
#> CHL2006 12.100971
#> CHN2006 14.764300
#> CMR2006  8.519357
#> COL2006  8.368768
#> CRI2006  9.068557
#> CYP2006  7.874743
#> DEU2006 14.001162
#> DNK2006 11.748412
#> ECU2006  7.551515
#> EGY2006 10.570508
#> ESP2006 13.127807
#> FIN2006 11.646010
#> FRA2006 13.502214
#> GBR2006 13.531254
#> GRC2006 11.174612
#> HKG2006 10.763297
#> HUN2006 10.592355
#> IDN2006 12.030365
#> IND2006 11.633171
#> IRL2006 11.543428
#> IRN2006 10.315344
#> ISL2006  7.042867
#> ISR2006 12.143375
#> ITA2006 13.284371
#> JOR2006  6.081093
#> JPN2006 14.450169
#> KEN2006  7.020711
#> KOR2006 13.598496
#> KWT2006  9.450534
#> LKA2006  9.174535
#> MAC2006  7.115140
#> MAR2006  9.132482
#> MEX2006 12.308697
#> MLT2006  9.197141
#> MMR2006  7.548097
#> MUS2006  7.805957
#> MWI2006  6.886646
#> MYS2006 13.560497
#> NER2006  6.121293
#> NGA2006  8.393132
#> NLD2006 12.816793
#> NOR2006 10.569590
#> NPL2006  6.871619
#> PAN2006  9.830518
#> PHL2006 10.892913
#> POL2006 11.339737
#> PRT2006 11.163732
#> QAT2006  9.479591
#> ROM2006 10.676721
#> SEN2006  7.170802
#> SGP2006 12.699639
#> SWE2006 12.701439
#> THA2006 12.584360
#> TTO2006  9.241514
#> TUN2006  9.466146
#> TUR2006 12.473791
#> TZA2006 10.205870
#> URY2006  8.572133
#> USA2006 14.406648
#> ZAF2006 11.696636
#> 
#> $imp_year
#>           imp_year
#> ARG2006  10.512190
#> AUS2006  12.779712
#> AUT2006  11.356127
#> BEL2006  11.837730
#> BGR2006   9.987278
#> BOL2006   9.357850
#> BRA2006  12.305863
#> CAN2006  12.409845
#> CHE2006  11.493759
#> CHL2006  11.281660
#> CHN2006  14.495644
#> CMR2006   7.730967
#> COL2006  10.567232
#> CRI2006  10.728940
#> CYP2006  10.351493
#> DEU2006  13.998173
#> DNK2006  10.921236
#> ECU2006   9.461344
#> EGY2006 -11.049074
#> ESP2006  12.527030
#> FIN2006  11.342308
#> FRA2006  12.732080
#> GBR2006  12.978547
#> GRC2006  10.472543
#> HKG2006  12.046917
#> HUN2006  11.065530
#> IDN2006  11.772378
#> IND2006  12.720817
#> IRL2006  10.667531
#> IRN2006  10.524743
#> ISL2006   8.538590
#> ISR2006  10.792297
#> ITA2006  12.770506
#> JOR2006   9.053934
#> JPN2006  12.975372
#> KEN2006   9.488157
#> KOR2006  12.211641
#> KWT2006  10.209792
#> LKA2006  10.324019
#> MAC2006   7.780500
#> MAR2006  10.159951
#> MEX2006  12.902123
#> MLT2006  10.231103
#> MMR2006   6.565826
#> MUS2006   9.466821
#> MWI2006   7.661901
#> MYS2006  10.012227
#> NER2006   7.450813
#> NGA2006   9.947689
#> NLD2006  12.188180
#> NOR2006  11.339707
#> NPL2006   6.171987
#> PAN2006   9.656596
#> PHL2006   8.474044
#> POL2006  11.381463
#> PRT2006  11.476783
#> QAT2006  10.247446
#> ROM2006  11.718380
#> SEN2006   7.714490
#> SGP2006  12.821630
#> SWE2006  12.006765
#> THA2006  11.429475
#> TTO2006   8.818622
#> TUN2006   9.394718
#> TUR2006  10.701788
#> TZA2006   8.825459
#> URY2006   8.426314
#> USA2006  15.364095
#> ZAF2006  12.120288
#>