Recover the estimates of the fixed effects after fitting (G)LMs

The system might not have a unique solution since we do not take collinearity into account. If the solution is not unique, an estimable function has to be applied to our solution to get meaningful estimates of the fixed effects.

Usage

fixed_effects(object = NULL, alpha_tol = 1e-08)

Arguments

object

an object of class "feglm".

alpha_tol

tolerance level for the stopping condition. The algorithm is stopped at iteration \(i\) if \(||\boldsymbol{\alpha}_{i} - \boldsymbol{\alpha}_{i - 1}||_{2} < tol ||\boldsymbol{\alpha}_{i - 1}|| {2}\). Default is 1.0e-08.

Value

A named list containing named vectors of estimated fixed effects.

References

Stammann, A. (2018). "Fast and Feasible Estimation of Generalized Linear Models with High-Dimensional k-way Fixed Effects". ArXiv e-prints.

Gaure, S. (n. d.). "Multicollinearity, identification, and estimable functions". Unpublished.

See also

Examples

# check the feglm examples for the details about clustered standard errors

# subset trade flows to avoid fitting time warnings during check
set.seed(123)
trade_2006 <- trade_panel[trade_panel$year == 2006, ]
trade_2006 <- trade_2006[sample(nrow(trade_2006), 500), ]

mod <- fepoisson(
  trade ~ log_dist + lang + cntg + clny | exp_year + imp_year,
  trade_2006
)

fixed_effects(mod)
#> $exp_year
#>          exp_year
#> ARG2006 12.354161
#> AUS2006 12.295867
#> AUT2006 11.460262
#> BEL2006 12.602932
#> BGR2006  9.955023
#> BOL2006  7.996867
#> BRA2006 13.196359
#> CAN2006 13.219168
#> CHE2006 11.505632
#> CHL2006 12.118175
#> CHN2006 14.781521
#> CMR2006  8.533977
#> COL2006  8.307502
#> CRI2006  9.084894
#> CYP2006  7.888434
#> DEU2006 14.018366
#> DNK2006 11.766103
#> ECU2006  7.567544
#> EGY2006 10.588030
#> ESP2006 13.144637
#> FIN2006 11.663078
#> FRA2006 13.519162
#> GBR2006 13.547807
#> GRC2006 11.192210
#> HKG2006 10.781007
#> HUN2006 10.610295
#> IDN2006 12.047878
#> IND2006 11.650341
#> IRL2006 11.558286
#> IRN2006 10.332926
#> ISL2006  6.950027
#> ISR2006 12.161167
#> ITA2006 13.301554
#> JOR2006  6.006858
#> JPN2006 14.467500
#> KEN2006  6.387530
#> KOR2006 13.615636
#> KWT2006  9.467513
#> LKA2006  9.189920
#> MAC2006  7.106948
#> MAR2006  9.149039
#> MEX2006 12.325664
#> MLT2006  9.158908
#> MMR2006  7.133583
#> MUS2006  7.811741
#> MWI2006  6.898888
#> MYS2006 13.576996
#> NER2006  5.288208
#> NGA2006  8.405552
#> NLD2006 12.834436
#> NOR2006 10.587597
#> NPL2006  6.528159
#> PAN2006  9.847054
#> PHL2006 10.909143
#> POL2006 11.357364
#> PRT2006 11.181306
#> QAT2006  9.494971
#> ROM2006 10.693677
#> SEN2006  5.616111
#> SGP2006 12.716678
#> SWE2006 12.719729
#> THA2006 12.601648
#> TTO2006  9.258660
#> TUN2006  9.482814
#> TUR2006 12.491000
#> TZA2006 10.222729
#> URY2006  8.530859
#> USA2006 14.423542
#> ZAF2006 11.712998
#> 
#> $imp_year
#>             imp_year
#> ARG2006 -0.165041214
#> AUS2006  2.102442310
#> AUT2006  0.679548946
#> BEL2006  1.160304051
#> BGR2006 -0.689523517
#> BOL2006 -1.319689007
#> BRA2006  1.628517578
#> CAN2006  1.731984696
#> CHE2006  0.816133839
#> CHL2006  0.605432549
#> CHN2006  3.818128362
#> CMR2006 -2.948658927
#> COL2006 -0.110143248
#> CRI2006  0.052387651
#> CYP2006 -0.325455402
#> DEU2006  3.321304124
#> DNK2006  0.244991918
#> ECU2006 -1.215898096
#> EGY2006 -0.473380456
#> ESP2006  1.850218644
#> FIN2006  0.666174002
#> FRA2006  2.055375781
#> GBR2006  2.303021793
#> GRC2006 -0.204577987
#> HKG2006  1.369945343
#> HUN2006  0.388630562
#> IDN2006  1.095288634
#> IND2006  2.043749909
#> IRL2006 -0.009954044
#> IRN2006 -0.152509608
#> ISL2006 -2.138124954
#> ISR2006  0.115852017
#> ITA2006  2.093499997
#> JOR2006 -1.622488274
#> JPN2006  2.298284174
#> KEN2006 -1.189110006
#> KOR2006  1.534748372
#> KWT2006 -0.466809416
#> LKA2006 -0.353356093
#> MAC2006 -2.898691925
#> MAR2006 -0.516716684
#> MEX2006  2.224371515
#> MLT2006 -0.445499402
#> MMR2006 -4.373996116
#> MUS2006 -1.219733984
#> MWI2006 -3.934918314
#> MYS2006 -0.665906258
#> NER2006 -3.265414769
#> NGA2006 -0.729561885
#> NLD2006  1.512128278
#> NOR2006  0.662582338
#> NPL2006 -4.565215479
#> PAN2006 -1.020333516
#> PHL2006 -2.276455453
#> POL2006  0.705214458
#> PRT2006  0.800238861
#> QAT2006 -0.429519902
#> ROM2006  1.043138044
#> SEN2006 -2.962421442
#> SGP2006  2.144622470
#> SWE2006  1.329580578
#> THA2006  0.752780074
#> TTO2006 -1.858710618
#> TUN2006 -1.281777087
#> TUR2006  0.025231431
#> TZA2006 -1.852027365
#> URY2006 -2.251062977
#> USA2006  4.686314244
#> ZAF2006  1.442960229
#>