Recover the estimates of the fixed effects after fitting (G)LMs

The system might not have a unique solution since we do not take collinearity into account. If the solution is not unique, an estimable function has to be applied to our solution to get meaningful estimates of the fixed effects.

Usage

fixed_effects(object = NULL, control = NULL)

Arguments

object

an object of class "feglm".

control

a list of control parameters. If NULL, the default control parameters are used.

Value

A named list containing named vectors of estimated fixed effects.

References

Stammann, A. (2018). "Fast and Feasible Estimation of Generalized Linear Models with High-Dimensional k-way Fixed Effects". ArXiv e-prints.

Gaure, S. (n. d.). "Multicollinearity, identification, and estimable functions". Unpublished.

See also

Examples

# check the feglm examples for the details about clustered standard errors

# subset trade flows to avoid fitting time warnings during check
set.seed(123)
trade_2006 <- trade_panel[trade_panel$year == 2006, ]
trade_2006 <- trade_2006[sample(nrow(trade_2006), 500), ]

mod <- fepoisson(
  trade ~ log_dist + lang + cntg + clny | exp_year + imp_year,
  trade_2006
)

fixed_effects(mod)
#> $exp_year
#>          exp_year
#> ARG2006 12.354182
#> AUS2006 12.295884
#> AUT2006 11.460273
#> BEL2006 12.602952
#> BGR2006  9.955045
#> BOL2006  7.997049
#> BRA2006 13.196374
#> CAN2006 13.219189
#> CHE2006 11.505648
#> CHL2006 12.118195
#> CHN2006 14.781523
#> CMR2006  8.533992
#> COL2006  8.307516
#> CRI2006  9.084915
#> CYP2006  7.888455
#> DEU2006 14.018382
#> DNK2006 11.766126
#> ECU2006  7.567525
#> EGY2006 10.588043
#> ESP2006 13.144660
#> FIN2006 11.663089
#> FRA2006 13.519188
#> GBR2006 13.547812
#> GRC2006 11.192226
#> HKG2006 10.781026
#> HUN2006 10.610367
#> IDN2006 12.048030
#> IND2006 11.650354
#> IRL2006 11.558309
#> IRN2006 10.332930
#> ISL2006  6.950040
#> ISR2006 12.161183
#> ITA2006 13.301576
#> JOR2006  6.006875
#> JPN2006 14.467518
#> KEN2006  6.387647
#> KOR2006 13.615647
#> KWT2006  9.467535
#> LKA2006  9.189935
#> MAC2006  7.106973
#> MAR2006  9.149060
#> MEX2006 12.325680
#> MLT2006  9.158931
#> MMR2006  7.133606
#> MUS2006  7.811756
#> MWI2006  6.898977
#> MYS2006 13.577020
#> NER2006  5.288808
#> NGA2006  8.405615
#> NLD2006 12.834446
#> NOR2006 10.587771
#> NPL2006  6.528169
#> PAN2006  9.847075
#> PHL2006 10.909153
#> POL2006 11.357386
#> PRT2006 11.181328
#> QAT2006  9.495016
#> ROM2006 10.693694
#> SEN2006  5.642459
#> SGP2006 12.716657
#> SWE2006 12.719739
#> THA2006 12.601663
#> TTO2006  9.258673
#> TUN2006  9.482837
#> TUR2006 12.491022
#> TZA2006 10.222754
#> URY2006  8.530873
#> USA2006 14.423558
#> ZAF2006 11.713008
#> 
#> $imp_year
#>          imp_year
#> ARG2006 10.432662
#> AUS2006 12.700175
#> AUT2006 11.277260
#> BEL2006 11.758003
#> BGR2006  9.908195
#> BOL2006  9.278026
#> BRA2006 12.226219
#> CAN2006 12.329683
#> CHE2006 11.413831
#> CHL2006 11.203132
#> CHN2006 14.415852
#> CMR2006  7.649043
#> COL2006 10.487572
#> CRI2006 10.650095
#> CYP2006 10.272262
#> DEU2006 13.919015
#> DNK2006 10.842685
#> ECU2006  9.381792
#> EGY2006 10.124333
#> ESP2006 12.447915
#> FIN2006 11.263887
#> FRA2006 12.653080
#> GBR2006 12.900723
#> GRC2006 10.393134
#> HKG2006 11.967618
#> HUN2006 10.986346
#> IDN2006 11.693003
#> IND2006 12.641439
#> IRL2006 10.587774
#> IRN2006 10.445182
#> ISL2006  8.459600
#> ISR2006 10.713559
#> ITA2006 12.691213
#> JOR2006  8.975206
#> JPN2006 12.895979
#> KEN2006  9.408605
#> KOR2006 12.132441
#> KWT2006 10.130901
#> LKA2006 10.244334
#> MAC2006  7.699018
#> MAR2006 10.080996
#> MEX2006 12.822101
#> MLT2006 10.152218
#> MMR2006  6.223685
#> MUS2006  9.377959
#> MWI2006  6.663847
#> MYS2006  9.931775
#> NER2006  7.332284
#> NGA2006  9.868127
#> NLD2006 12.109837
#> NOR2006 11.260309
#> NPL2006  6.032495
#> PAN2006  9.577375
#> PHL2006  8.321245
#> POL2006 11.302907
#> PRT2006 11.397952
#> QAT2006 10.168192
#> ROM2006 11.640851
#> SEN2006  7.635280
#> SGP2006 12.742339
#> SWE2006 11.927184
#> THA2006 11.350503
#> TTO2006  8.739001
#> TUN2006  9.315938
#> TUR2006 10.622928
#> TZA2006  8.745702
#> URY2006  8.346655
#> USA2006 15.283867
#> ZAF2006 12.040666
#>