Covariance matrix for LMs

Covariance matrix for the estimator of the structural parameters from objects returned by felm. The covariance is computed from the hessian, the scores, or a combination of both after convergence.

Usage

# S3 method for class 'felm'
vcov(
  object,
  type = c("hessian", "outer.product", "sandwich", "clustered"),
  ...
)

Arguments

object

an object of class "felm".

type

the type of covariance estimate required. "hessian" refers to the inverse of the negative expected hessian after convergence and is the default option. "outer.product" is the outer-product-of-the-gradient estimator. "sandwich" is the sandwich estimator (sometimes also referred as robust estimator), and "clustered" computes a clustered covariance matrix given some cluster variables.

...

additional arguments.

Value

A named matrix of covariance estimates.

See also

Examples

# same as the example in felm but extracting the covariance matrix

# subset trade flows to avoid fitting time warnings during check
set.seed(123)
trade_2006 <- trade_panel[trade_panel$year == 2006, ]
trade_2006 <- trade_2006[sample(nrow(trade_2006), 500), ]

mod <- felm(
  trade ~ log_dist + lang + cntg + clny | exp_year + imp_year | pair,
  trade_2006
)

round(vcov(mod, type = "clustered"), 5)
#>           [,1]      [,2]      [,3]        [,4]
#> [1,]  323125.2 -133763.4   1337704    152765.3
#> [2,] -133763.4 1541792.7   1762259   -279416.5
#> [3,] 1337703.9 1762259.2  65752028 -10417337.9
#> [4,]  152765.3 -279416.5 -10417338  10023467.4